Rename and refactor OCR handling
Renamed `ocr.ts` to `processGameFrame.ts` and refactored OCR worker pool logic into `workOCR.ts`. Updated module imports across the codebase to reflect these changes, enhancing modularity and code organization.
